Add restore settings to options page
Previously the checkboxes on the options page would always be unchecked
when the options page was loaded. The user has to know which folders are
already being sorted before unchoosing them. In order to accomplish
this, I moved the binary search and comparison methods to a util file
and loaded it before the other files.