Change entire style
The default list and buttons given by HTML are not flexible enough for
interactive trees (see https://www.jstree.com/). From here, I went
through and replaced most elements with a DIV and custom class. For the
bookmark tree, flexbox allowed me to format the tree to my liking. I
added sass in order to allow extending and variables. After this, I
changed the javascript code in order to handle the new system.

A few minor changes were made to the remaining HTML and CSS.